backend/requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.3.1
aniso8601==9.0.1
appnope==0.1.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
ccxt==2.5.12
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
debugpy==1.5.1
decorator==5.1.1
dnspython==2.2.1
entrypoints==0.4
executing==0.8.3
Flask==2.2.2
Flask-RESTful==0.3.9
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.15.2
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter_client==7.4.8
jupyter_core==5.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.1
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==22.3.1
platformdirs==2.5.2
prompt-toolkit==3.0.36
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycares==4.3.0
pycparser==2.21
Pygments==2.11.2
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2022.7
pyzmq==23.2.0
redis==4.4.0
requests==2.28.1
rq==1.11.1
setuptools==65.5.0
six==1.16.0
stack-data==0.2.0
tornado==6.2
traitlets==5.7.1
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
yarl==1.8.2
zipp==3.11.0
